pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.1.0",
    "ddgs>=9.5.5",
    "fastapi[standard]>=0.118.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.77",
    "langchain-ollama>=0.3.9",
    "langchain-openai>=0.3.32",
    "langchain-qdrant>=0.2.1",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langmem>=0.0.29",
    "markdownify>=1.2.0",
    "nicegui>=3.0.4",
    "notion-client>=2.5.0",
    "numpy>=2.3.2",
    "ollama>=0.6.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg>=3.2.11",
    "pyabsa>=2.4.2",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.4",
    "pymysql>=1.1.2",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "readability-lxml>=0.8.4.1",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "selenium>=4.35.0",
    "sentence-transformers>=5.1.1",
    "websockets>=15.0.1",
]
